refactored models to keep code DRY with abstract entity
